requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
dotenv==0.9.9
pillow==11.3.0
psycopg==3.2.10
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
